5/4変更分

+++ app/Http/Controllers/PrefController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
 use App\mst_prefecture as Pref;
 
 class PrefController extends Controller
@@ -12,17 +11,26 @@ class PrefController extends Controller
         return view('list');
     }
 
-    public function update(Request $request){
-        return "kousin";
-    }
-
     public function edit(Request $request){
-        return view('edit');
+        $arr_request=$request->input();
+        return view('edit',compact('arr_request'));
     }
 
     public function confirm(Request $request){
         $arr_request=$request->input();
-        return view('confirm',compact('arr_request'));
+        if(!isset($arr_request['prefecture_cd']) && isset($arr_request['prefecture_name'])){
+            $errors[]='・地域コードが入力されていません！';
+            return view('edit',compact('arr_request','errors'));
+        }else if(isset($arr_request['prefecture_cd']) && !isset($arr_request['prefecture_name'])){
+            $errors[]='・地域名が入力されていません！';
+            return view('edit',compact('arr_request','errors'));
+        }else if(!isset($arr_request['prefecture_cd']) && !isset($arr_request['prefecture_name'])){
+            $errors[]='・地域コードが入力されていません！';
+            $errors[]='・地域名が入力されていません！';
+            return view('edit',compact('arr_request','errors'));
+        }else{
+            return view('confirm',compact('arr_request','errors'));
+        }
     }
 
     public function complete(Request $request){
@@ -38,11 +46,17 @@ class PrefController extends Controller
             }else{
                 return 'can’t insert!';
             }
-        }else{
+        }elseif($arr_request['process'] == 'delete'){
             if(Pref::where('prefecture_cd','=',$arr_request['prefecture_cd'])->delete() == 1){
                 return view('complete',compact('arr_request'));
             }else{
                 return 'can’t delete!';
+            }
+        }else{
+            if(Pref::where('prefecture_cd','=',$arr_request['prefecture_cd'])->update(['prefecture_name' => $arr_request['prefecture_name']]) == 1){
+                return view('complete',compact('arr_request'));
+            }else{
+                return 'can’t update!';
             }
         }
         
@@ -50,9 +64,12 @@ class PrefController extends Controller
 
     public function search(Request $request){
         $arr_request=$request->input();
+        if(isset($arr_request['prefecture_cd'])){
+            $arr_request['prefecture_cd']=str_pad($arr_request['prefecture_cd'], 2, 0, STR_PAD_LEFT);
+        }
         $db=Pref::get();
-        if(!isset($arr_request['prefecture_cd']) && isset($arr_request['prefecture_cd'])){
-            $data=Pref::where('prefecture_name','like',$arr_request['prefecture_name'])
+        if(!isset($arr_request['prefecture_cd']) && isset($arr_request['prefecture_name'])){
+            $data=Pref::where('prefecture_name','like','%'.$arr_request['prefecture_name'].'%')
             ->paginate(10);
             $query=array(
                 'prefecture_cd' => '',
@@ -73,13 +90,13 @@ class PrefController extends Controller
             );
         }else{
             $data=Pref::where('prefecture_cd','=',$arr_request['prefecture_cd'])
-            ->where('prefecture_name','like',$arr_request['prefecture_name'])
+            ->where('prefecture_name','like','%'.$arr_request['prefecture_name'].'%')
             ->paginate(10);
             $query=array(
                 'prefecture_cd' => $request->input('prefecture_cd'),
                 'prefecture_name' => $request->input('prefecture_name')
             );
         }
-        return view('search',compact('data','request','db','query'));
+        return view('search',compact('data','arr_request','db','query'));
     }
 }
